- renames references to large file upload tasks
